src/core/utils.py
"""
Assignment 11 — Helper Utilities
"""
from google.genai import types


async def chat_with_agent(
    agent,
    runner,
    user_message: str,
    session_id: str | None = None,
    user_id: str = "student",
) -> tuple[str, object]:
    """Send a message to an ADK agent and return the response text.

    Args:
        agent:        The LlmAgent instance.
        runner:       The InMemoryRunner instance.
        user_message: Plain text message to send.
        session_id:   Optional session ID to continue an existing conversation.
        user_id:      User identifier (used by rate limiter & session anomaly plugins).

    Returns:
        Tuple of (response_text, session).
    """
    app_name = runner.app_name

    session = None
    if session_id is not None:
        try:
            session = await runner.session_service.get_session(
                app_name=app_name, user_id=user_id, session_id=session_id
            )
        except (ValueError, KeyError):
            pass

    if session is None:
        session = await runner.session_service.create_session(
            app_name=app_name, user_id=user_id
        )

    content = types.Content(
        role="user",
        parts=[types.Part.from_text(text=user_message)],
    )

    final_response = ""
    async for event in runner.run_async(
        user_id=user_id, session_id=session.id, new_message=content
    ):
        if hasattr(event, "content") and event.content and event.content.parts:
            for part in event.content.parts:
                if hasattr(part, "text") and part.text:
                    final_response += part.text

    return final_response, session
