requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
nemoguardrails>=0.10.0
openai>=1.30.0
litellm>=1.40.0
python-dotenv>=1.0.0
streamlit>=1.35.0
